feat: add health check and stock update listener
